Add files via upload
Created restart.py, used to call the main script. This should hopefully help with the exception produced when trying to manually update the bot. calling the main script from within the main script, unsurprisingly, seems to be the cause of the issue. Also, i'm dumb and messed up the last commit I am experimenting with github pulls